refactor code to use new launchpad-auth package
