+++ packages/api/package.json
@@ -49,11 +49,13 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@cumulus/aws-client": "1.21.0",
+    "@cumulus/collection-config-store": "1.20.0-alpha1",
     "@cumulus/cmr-client": "1.21.0",
     "@cumulus/cmrjs": "1.21.0",
     "@cumulus/common": "1.21.0",
     "@cumulus/errors": "1.21.0",
     "@cumulus/ingest": "1.21.0",
+    "@cumulus/launchpad-auth": "1.20.0-alpha1",
     "@cumulus/logger": "1.21.0",
     "@cumulus/message": "1.21.0-alpha3",
     "@cumulus/pvl": "1.21.0",
