+++ packages/api/tests/app/test-launchpadAuth.js
@@ -4,7 +4,7 @@ const test = require('ava');
 const sinon = require('sinon');
 const request = require('supertest');
 const { secretsManager } = require('@cumulus/aws-client/services');
-const launchpad = require('@cumulus/common/launchpad');
+const launchpad = require('@cumulus/launchpad-auth');
 const { randomId } = require('@cumulus/common/test-utils');
 const EsCollection = require('../../es/collections');
 const models = require('../../models');
