+++ packages/api/tests/endpoints/test-granules.js
@@ -21,7 +21,7 @@ const { CMR } = require('@cumulus/cmr-client');
 const {
   metadataObjectFromCMRFile
 } = require('@cumulus/cmrjs/cmr-utils');
-const launchpad = require('@cumulus/common/launchpad');
+const launchpad = require('@cumulus/launchpad-auth');
 const { randomString, randomId } = require('@cumulus/common/test-utils');
 
 const assertions = require('../../lib/assertions');
