+++ packages/launchpad-auth/index.js
