+++ tasks/post-to-cmr/package.json
@@ -41,6 +41,7 @@
     "@cumulus/common": "1.21.0",
     "@cumulus/cumulus-message-adapter-js": "^1.1.0",
     "@cumulus/errors": "1.21.0",
+    "@cumulus/launchpad-auth": "1.20.0-alpha1",
     "lodash.keyby": "^4.6.0"
   },
   "devDependencies": {
